Fixes #15170 - Moving timeout to a large enough value
This is necessary because we have various operations in Katello that
are synchronus, especially around our interaction with candlepin and
manifest import that we will often time-out during operation.

We need to keep this timeout value high until the API calls we rely on
are implemented in an async manner.