Fixes #16008 - allow setting puppet_repo_root
Previously, this setting was not set, which made it default to
`/etc/puppet/environments` in Katello.

This patch adds an explicit setting for `puppet_repo_root`, but
defaults to `/etc/puppet/environments/` on puppet 3 systems, and
`/etc/puppetlabs/code/environments/` for puppet 4 systems.

Additionally, the "no-80chars-check" was removed, since the flag no longer
exists in puppet-lint 2.y.