Improve ownership of config files
At some point the bundler file was ensured to be owned by the katello
user. This traces back to the start of the module
(ae7d4aa1039445bbf130d9951ab36d866d551a3b) but there is no real reason
for this. We could not manage the file at all, but by ensuring it's
owned by root we improve security for existing users.

We also tighten the ownership on the katello.yaml config file.