Refactor module
This rewrites large parts of the module. The main goal is to split
init.pp into several classes that each describe a part of the stack.
This is then composed in init.pp where additional chaining is applied.
While it doesn't add parameters to choose the composition, it does open
the path to it.

It also relies more on using puppet-foreman. This means $config_dir is
removed in favor of $foreman::plugin_config_dir. Additionally it uses
Foreman::rake as chaining rather than the internal Execs. Similarly it
now notifies the tasks instead of the Service.

All classes use explicit parameters in the local scope and have their
own specs.