Simplify package name handling
f614db22fdb78f67b2e778f1729512683de4a251 removed the use of
$rubygem_katello_ostree variable, but not its definition.

We also don't want to hard fail because we sometimes want to compile on
other distros. The most common use case is CI testing our installer by
running it with --help. Since that happens on Travis (which is Ubuntu)
it hard fails.