Redesign classes to allow split deployments
Every class that provides an aspect of the deployment is now standalone.
This means:

* candlepin
* pulp
* qpid
* application

They can just be included and work by themselves. All class parameters
are tunables that users will commonly modify.

There's also shared parameters which are common among the standalone
classes. These are meant to be changed via katello::params. Either via
Hiera or by including them in a profile class before the instance
classes themselves.

The katello::globals class has globals that a user via the installer
might want to change. The katello::params class is not intended to be
exposed via the installer.

The main katello class remains as a compatibility class. In the future
this should probably be converted into a role but perhaps it is no
needed at all.