Refs #29039 - Use foreman's Apache config standalone (#324)
* Refs #29039 - Use foreman's Apache config standalone

By using only the Apache config, it's possible to use the Apache config
without Foreman installed.

* Move the repo_export_dir to application.pp

The repo export directory default matches with the
pulp_export_destination setting. This means it's only needed on a
Katello application server, not on a Pulp server.

By creating an anchor, we can collect this if needed and otherwise fall
back to a simple mkdir -p. By ordering these it should be idempotent,
guarantee the correct permissions on /var/lib/pulp and the
katello-export subdirectory.