Move Artemis client DN construction to application
The Candlepin server should not deploy the Foreman client certificates.
Instead, it should be passed in as a parameter. This allows a standalone
deployment. In case an all-in-one deployment is done the value is taken
from the application.

The Candlepin server certificates are no longer needed on the
application side: the Foreman certificates are used as client
certificates.