Calculate the client DN for Artemis, ignoring empty values (#457)
* Calculate the client DN for Artemis, ignoring empty values

Moving the computation of the Artemis client DN into this class
instead of relying on the value provided directly by classes in
puppet-certs aims to solve two problems:

 1) Testing at all levels of the value of the client DN
 2) Adding a function to calculate the DN based on a set of values and
    ignoring any empty values which do not end up in the actual certificate

The Candlepin server should not deploy the Foreman client certificates.
Instead, it should be passed in as a parameter. This allows a standalone
deployment. In case an all-in-one deployment is done the value is taken
from the application.

The Candlepin server certificates are no longer needed on the
application side: the Foreman certificates are used as client
certificates.

Co-authored-by: Ewoud Kohl van Wijngaarden <ewoud@kohlvanwijngaarden.nl>